refactor: Rename utils into data
